Extract PTY engine (#49)
