--- .opencode/plans/1782520155016-cosmic-garden.md
@@ -1,79 +1,0 @@
-# Fix: PanelComponent as empty container with composable child views
-
-## Problem
-`PanelComponent` currently renders everything (tabs, menu, hints, info) in a single monolithic `render()` call. This makes it impossible to use separate `top_panel`/`bottom_panel` views — each instance renders the full thing, causing double-render.
-
-## Solution
-Make `PanelComponent` an empty container/shell. It provides the background area and delegates to child components that each render a specific feature.
-
-### New types
-
-```rust
-/// A single row/segment within a panel. Each renders independently.
-pub trait PanelSegment<Id: Copy + Eq + Ord>: std::fmt::Debug {
-    fn render(&mut self, frame: &mut UiFrame, area: Rect, config: &PanelConfig);
-}
-```
-
-Wait — actually, the simplest approach is probably just having `PanelComponent` hold a `Vec<Box<dyn Panel<WindowId<Id>>>>` as children, and rendering each one into its area. But `PanelTrait::render` signature is complex (14 params). The child components would need the full context.
-
-Actually, rethinking this based on what the user said — "panel component should, by itself, render an empty panel. you fill it in with composition":
-
-The simplest interpretation: `PanelComponent` is a layout container that positions children. Children implement `PanelTrait` (the existing trait). `PanelComponent::render()`:
-1. Fills background
-2. Calls each child's `render()` in order
-3. Each child renders its own content into its allocated area
-
-`PanelConfig` changes: instead of just flags, it holds the actual child views that PanelComponent composes.
-
-```rust
-pub struct PanelConfig<Id: Copy + Eq + Ord> {
-    /// Optional top-bar children — each is rendered in the top panel area
-    pub top: Vec<Box<dyn Panel<Id>>>,
-    /// Optional bottom-bar children — each is rendered in the bottom panel area  
-    pub bottom: Vec<Box<dyn Panel<Id>>>,
-}
-```
-
-### How existing PanelComponent gets split
-Break the 878-line `PanelComponent` into small focused components:
-- `WindowTabs` — renders clickable window tabs
-- `MenuIcon` — renders hamburger "≡ appname"
-- `StatusLine` — renders WM mode messages
-- `Indicators` — renders selection/copy/mouse/clipboard toggles
-- `KeybindingHints` — renders bottom keybinding shortcuts
-- `InfoBar` — renders hostname/version/platform
-- `PanelBg` — fills background with theme color
-
-Each is a standalone struct implementing `PanelTrait`.
-
-### PanelComponent itself
-Becomes a thin container. It holds `PanelConfig` internally.
-On `render()`: fills background, then iterates `top` children and `bottom` children, calling each one's `render()`.
-
-### end result
-- `new_standalone` fills top: [WindowTabs, MenuIcon, StatusLine, Indicators], bottom: [KeybindingHints, InfoBar]
-- `new_embedded` fills top: [], bottom: [KeybindingHints]
-- No double-render — each child renders only its own content
-- PanelConfig holds the actual child views, not just flags
-
-### Migration
-- Keep existing `PanelComponent` code but extract each feature into its own type
-- This is primarily a reorganization of existing code, not rewriting
-- All tests adapt to new type structure
-
-### files to modify
-- `crates/term-wm-core/src/panel_config.rs` — add view fields to PanelConfig
-- `crates/term-wm-core/src/panel_trait.rs` — no changes needed
-- `crates/term-wm-ui-components/src/panel.rs` — refactor into child types + container
-- `crates/term-wm-core/src/window/window_manager/mod.rs` — use PanelConfig children
-- `term-wm/src/lib.rs` — compose appropriate children for standalone/embedded
-- `src/main.rs` — compose appropriate children
-
-### verification
-```bash
-cargo build --workspace
-cargo clippy --workspace --all-targets --all-features -- -D warnings
-cargo test
-```
-ArgTuner compiles. No double-render. Embedded mode: bottom hints only. Standalone mode: all features.
