+++ crates/term-wm-core/src/io/mod.rs
@@ -1,6 +1,4 @@
-pub mod clipboard;
 pub mod console;
-pub mod title;
 pub mod utils;
 
 pub mod event_source;
