+++ crates/term-wm-core/src/lib.rs
@@ -10,8 +10,11 @@ pub mod event_loop;
 pub mod io;
 pub mod keybindings;
 pub mod layout;
-pub mod pane;
-pub mod pty;
+pub use term_wm_pty_engine::Pane;
+pub use term_wm_pty_engine::Pty;
+pub use term_wm_pty_engine::PtyResult;
+pub use term_wm_pty_engine::clipboard;
+pub use term_wm_pty_engine::title;
 pub mod runner;
 pub mod theme;
 pub mod top_panel_trait;
