+++ crates/term-wm-pty-engine/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod clipboard;
+pub mod pane;
+pub mod pty;
+pub mod title;
+
+pub use pane::Pane;
+pub use pty::{Pty, PtyResult};
